created visualizations class
