+++ pyautoml/visualizations/util.py
@@ -1,0 +1,14 @@
+from pyautoml.config import cfg, DEFAULT_IMAGE_DIR
+from pyautoml.util import _make_dir
+
+
+def _make_image_dir():  # pragma: no cover
+
+    if not cfg["images"]["dir"]:
+        image_dir = DEFAULT_IMAGE_DIR
+    else:
+        image_dir = cfg["images"]["dir"]
+
+    _make_dir(image_dir)
+
+    return image_dir
